+++ backend/rest_api/views/__init__.py
@@ -1,0 +1,3 @@
+from .user import *
+from .article import *
+from .community import *
