+++ backend/rest_api/views/community.py
@@ -1,0 +1,62 @@
+from ..models import Community
+from ..serializer import  CommunitySerializer, CommunityUsersSerializer
+from rest_framework.views import APIView
+from rest_framework import status
+from django.http import Http404
+from rest_framework.response import Response
+from rest_framework.generics import get_object_or_404
+from rest_framework.views import APIView
+from rest_framework.permissions import AllowAny
+
+
+
+class CommunityMembersAPIView(APIView):
+    authentication_classes = []
+    permission_classes = [AllowAny, ]
+
+    def get(self, request, pk, *args, **kwargs):
+        communities = get_object_or_404(Community, pk=pk)
+        serializer = CommunityUsersSerializer(communities)
+        return Response(
+            serializer.data,
+        )
+
+
+class CommunityList(APIView):
+    def get(self,request, format=None):
+        communities = Community.objects.all()
+        selializer = CommunitySerializer(communities, many=True)
+        return Response(selializer.data)
+    
+    def post(self, request, format=None):
+        serializer = CommunitySerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class CommunityDetail(APIView):
+    def get_object(self, pk):
+        try:
+            return Community.objects.get(pk=pk)
+        except Community.DoesNotExist:
+            raise Http404
+
+    def get(self, request, pk, format=None):
+        snippet = self.get_object(pk)
+        serializer = CommunitySerializer(snippet)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        snippet = self.get_object(pk)
+        serializer = CommunitySerializer(snippet, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        snippet = self.get_object(pk)
+        snippet.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
